[SofaQtQuickGUI] Refactor SofaScene to move the rendering out of it (breaking)
