+++ applications-dev/plugins/SofaQtQuickGUI/src/SofaQtQuickGUI/SofaScene.h
@@ -69,7 +69,7 @@ class SOFA_SOFAQTQUICKGUI_API SofaScene : public QObject, private sofa::simulati
     friend class EditView;
 
     friend class PickUsingRasterizationWorker;
-    friend bool LoaderProcess(SofaScene* scene, QOffscreenSurface* surface);
+    friend bool LoaderProcess(SofaScene* scene);
 
 public:
     explicit SofaScene(QObject *parent = 0);
@@ -235,12 +235,14 @@ public:
     const sofa::simulation::Node::SPtr& sofaRootNode() const {return mySofaRootNode;}
     sofa::simulation::Node::SPtr& sofaRootNode() {return mySofaRootNode;}
 
+    void prepareSceneForDrawing() ;
+
 protected:
     /// \brief      Low-level drawing function
     /// \attention  Require an opengl context bound to a surface, viewport / projection / modelview must have been set
     /// \note       The best way to display a 'Scene' is to use a 'Viewer' instead of directly call this function
-    void drawEditorView(const SofaViewer& viewer, const QList<SofaComponent*>& roots = QList<SofaComponent*>(),
-              bool doDrawSelected=true, bool doDrawManipulator=true) const;
+    //void drawEditorView(const SofaViewer& viewer, const QList<SofaComponent*>& roots = QList<SofaComponent*>(),
+    //          bool doDrawSelected=true, bool doDrawManipulator=true) const;
 
     /// \brief      Low-level function for mechanical state particle picking
     /// \note       The best way to pick a particle is to use a Viewer instead of directly call this function
@@ -251,18 +253,17 @@ protected:
     /// \attention  Require an opengl context bound to a surface, viewport / projection / modelview must have been set
     /// \note       The best way to pick an object is to use a Viewer instead of directly call this function
     /// \return     A 'Selectable' containing the picked object
-    Selectable* pickObject(const SofaViewer& viewer, const QPointF& ssPoint, const QStringList& tagList, const QList<SofaComponent*>& roots = QList<SofaComponent*>());
-
-    void drawSelectedComponents(sofa::core::visual::VisualParams *visualParams) const ;
-    void drawManipulator(const SofaViewer& viewer) const ;
+    //Selectable* pickObject(const SofaViewer& viewer, const QPointF& ssPoint, const QStringList& tagList, const QList<SofaComponent*>& roots = QList<SofaComponent*>());
+    //void drawSelectedComponents(sofa::core::visual::VisualParams *visualParams) const ;
+    //void drawManipulator(const SofaViewer& viewer) const ;
 
     /// This is calling the drawVisual() method on each Sofa object in the scene
-    void drawVisuals(const SofaViewer& viewer) const ;
+    //void drawVisuals(const SofaViewer& viewer) const ;
 
     /// This is calling the draw() method on each sofa object in the scene
-    void drawDebugVisuals(const SofaViewer& viewer) const ;
+    //void drawDebugVisuals(const SofaViewer& viewer) const ;
 
-    sofa::core::visual::VisualParams* setupVisualParams(sofa::core::visual::VisualParams*) const;
+    //sofa::core::visual::VisualParams* setupVisualParams(sofa::core::visual::VisualParams*) const;
     void setupCamera(int width, int height, const SofaViewer& viewer) const ;
     void clearBuffers(const QSize &size, const QColor& color, const QImage& image=QImage()) const ;
 
@@ -279,7 +280,8 @@ private:
     QUrl                                        mySourceQML;
     QString                                     myPath;
     QString                                     myPathQML;
-    mutable bool                                myVisualDirty;
+    mutable bool                                myVisualDirty {true} ;
+    mutable bool                                myTextureAreDirty {true} ;
     double                                      myDt;
     bool                                        myAnimate;
     bool                                        myDefaultAnimate;
@@ -294,11 +296,6 @@ private:
     QList<Manipulator*>                         myManipulators;
     Manipulator*                                mySelectedManipulator;
     SofaComponent*                              mySelectedComponent;
-
-    //TODO(dmarchal) this shouldn't be in a SofaScene class. SofaApplication sound a better place ?
-    QOpenGLShaderProgram*                       myHighlightShaderProgram;
-    QOpenGLShaderProgram*                       myPickingShaderProgram;
-    QOpenGLFramebufferObject*                   myPickingFBO;
 };
 
 }
